separate baselayout components
