+++ src/Scenes/MainPage/index.tsx
@@ -1,48 +1,17 @@
-import { Card, Grid } from "@material-ui/core";
 import React from "react";
 import BaseLayout from "Scenes/components/BaseLayout";
-import Spacer from "Scenes/components/Spacer";
-
-// this component is only a placeholder and should be removed after actual components have been replaced.
-const Filler = ({ text, height }: {text:string, height: string}) => (
-  <Card 
-    style={{
-      backgroundColor: 'gainsboro',
-      height,
-    }}
-    elevation={0}
-  >
-    {text}
-  </Card>
-);
+import HomeCenterSection from "./components/HomeCenterSection";
+import HomeLeftSidebar from "./components/HomeLeftSidebar";
+import HomeRightSidebar from "./components/HomeRightSidebar";
 
 const Index = () => {
-
-  const centerComponent = () => (
-    <Grid container direction="column" wrap="nowrap">
-       <Grid item>
-         <Filler height="100px" text="insert create post component here"/>
-       </Grid>
-       <Spacer spacing={2} orientation="h"/>
-       <Grid item xs>
-         <Filler height="700px" text="insert filter and message list component here"/>
-       </Grid>
-     </Grid>
-  );
-  const rightComponent = () => (
-    <Filler height="100%" text="insert right sidebar component here"/>
-  );
-  const leftComponent = () => (
-    <Filler height="100%" text="insert left sidebar component here"/>
-  );
-
   return (
     <BaseLayout
-      CenterComponent={centerComponent}
-      RightComponent={rightComponent}
-      LeftComponent={leftComponent}
+      CenterComponent={HomeCenterSection}
+      RightComponent={HomeRightSidebar}
+      LeftComponent={HomeLeftSidebar}
     />
   );
 };
 
-export default Index;
+export default Index;
